pulling told tests into their own module, resetting app order to avoid South breaking in test suite
