+++ mwana/apps/smgl/tests/__init__.py
@@ -7,5 +7,6 @@ from .test_lookups import *
 from .test_models import *
 from .test_pregnancy import *
 from .test_refer import *
+from .test_told import *
 from .test_utils import *
 from .test_views import *
